#1143 added class for option modes
